Fixed user administration.

+++ minimark/src/main/webapp/WEB-INF/jsp/admin/users/list.jsp
@@ -1,36 +1,20 @@
-<%@ include file="../../common.jspf" %>
-<!DOCTYPE html PUBLIC "-//W3C//DTD XHTML 1.0 Strict//EN" "http://www.w3.org/TR/xhtml1/DTD/xhtml1-strict.dtd">
-<html xmlns="http://www.w3.org/1999/xhtml">
-    <head>
-        <%@ include file="../../head.jspf" %>
-    </head>
-    <body>
-        <div id="nonFooter">            
-            <jsp:include page="../../header.jsp"/>
-            <div id="content"> 
-                <div id="content_main">
-                    <h1>Users</h1>
-                    
-                    <table>
-                        <c:forEach var="user" items="${users}">
-                            <tr>
-                                <td>${user.username}</td>
-                                <td>
-                                    <a href="edit.form?id=${user.id}">Edit</a>
-                                    <a href="delete.html?id=${user.id}">Delete</a>
-                                </td>
-                            </tr>
-                        </c:forEach>
-                    </table>
-                    <a href="create.html">New</a>                            
-                    
-                    <c:if test="${empty users}">
-                        No users in the DB
-                    </c:if>
-                </div>
-                <jsp:include page="../../menu.jsp"/>
-            </div>
-        </div>
-        <jsp:include page="../../footer.jsp"/>        
-    </body>
-</html>
+<%@ include file="/WEB-INF/jsp/common.jspf" %>
+<h3>Users</h3>
+
+<div>
+    <a class="action-add" href="create.html"><spring:message code="AddUser" text="?AddUser?"/></a>
+</div>
+
+<div class="displaytag">
+    <display:table  id="user"  name="users" sort="list" pagesize="20" defaultsort="1" defaultorder="ascending" >
+        <display:column property="user.username" titleKey="Username" class="left-aligned"/>
+        <display:column property="name" titleKey="Name" class="left-aligned"/>
+        <display:column property="email" titleKey="Email" class="left-aligned" autolink="true"/>
+        <display:column property="site" titleKey="Site" class="left-aligned" autolink="true"/>
+        <display:column>
+            <a class="action-edit" href="edit.form?id=${user.user.id}"><spring:message code="Edit" text="?Edit?"/></a>
+            <spring:message code='confirmDeleteUser' text='?confirmDeleteUser?' var="confirmDeleteUserMessage" arguments="${user.user.username}" javaScriptEscape="true"/>
+            <a class="action-delete" href="delete.html?id=${user.user.id}" onclick="return confirm('${confirmDeleteUserMessage}')"><spring:message code="Delete" text="?Delete?"/></a>
+        </display:column>
+    </display:table>
+</div>
